use TIMESTAMP for Disk Forensics This patch converts pyflag to using TIMESTAMP database fields for time fields. (formerly using int). It updates dbtool and all disk forensics reports to use the new format.
This provides a number of advantages:
1. tables in reports are simplified, they just have to select the column
(eg. select atime) rather than applying a function (eg. select from_unixtime(atime))

2. filtering on time fields is now MUCH faster because a function is not applied
to the column and the index can be used.

3. times are still stored internally in UTC in mysql. They are displayed by default
in timezone of the analysis system. If we wish to display them in another timezone
this is as easy as setting time_zone variable in mysql before the select.