MSN Changes
Tried to add some actual test cases to the MSN stuff. There were a couple of bugs. I think this still needs work, in particular the file transfer report isn't quite right. Hopefully this will be a reasonable start. We need more tests. I will try and make use of the std captures and add them in. At least now it seems to be working.